[🔨refactor] 가격 형식 변환 함수 분리 #85
